Add test: payments
